Revert "Improve docs breadcrumbs (#2873)"
This reverts commit d784842a31aec38a9cba8faf52877f1f2cd6bcfc.
