+++ docs/public/feed.rss
@@ -5,7 +5,7 @@
     <link>https://docs.xmlui.org</link>
     <description>Latest updates, tutorials, and insights for XMLUI - the declarative UI framework. Stay informed about new features, best practices, and community highlights.</description>
     <language>en</language>
-    <lastBuildDate>Mon, 23 Feb 2026 13:32:35 GMT</lastBuildDate>
+    <lastBuildDate>Wed, 21 Jan 2026 13:35:12 GMT</lastBuildDate>
 
   </channel>
 </rss>
