+++ xmlui/src/components-core/script-runner/eval-tree-sync.ts
@@ -619,7 +619,7 @@ function evalNewExpression(
   thread: LogicalThread,
 ): any {
   // --- Allowed constructors
-  const allowedConstructors = new Map<string, new (...args: any[]) => any>([
+  const allowedConstructors = new Map([
     ["String", String],
     ["Date", Date],
     ["Blob", Blob],
