+++ xmlui/src/components/Table/TableNative.tsx
@@ -8,6 +8,7 @@ import type {
   HeaderContext,
   PaginationState,
   Row,
+  RowData,
 } from "@tanstack/react-table";
 import {
   flexRender,
@@ -57,6 +58,24 @@ import {
 
 // =====================================================================================================================
 // Helper types
+
+// --- Declaration merging, see here: https://tanstack.com/table/v8/docs/api/core/table#meta
+declare module "@tanstack/table-core" {
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  interface TableMeta<TData extends RowData> {
+    cellRenderer: (...args: any[]) => any;
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  interface ColumnMeta<TData extends RowData, TValue> {
+    style?: CSSProperties;
+    className?: string;
+    starSizedWidth?: string;
+    accessorKey?: string;
+    pinTo?: string;
+    cellRenderer?: (row: any, rowIdx: number, colIdx: number, value?: any) => ReactNode;
+  }
+}
 
 /**
  * This type describes an arbitraty table row that has an integer identifier and an order index.
